# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DM Interactive Table API
  version: 2.0.0-alpha.23
  description: API HTTP para sesiones, campañas, escenas, mapas, combate, dados e IA.
servers:
  - url: http://localhost:4000/api
security:
  - bearerAuth: []
  - dmCookie: []
  - tableCookie: []
paths:
...
  /table-access:
    get:
      summary: Consultar el acceso temporal de una campaña
      operationId: getTableAccess
      parameters:
        - name: campaignId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Estado del acceso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableAccessStatus"
    post:
      summary: Generar un código temporal de mesa
      operationId: createTableAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [campaignId]
              properties:
                campaignId:
                  type: string
                sessionId:
                  type: string
                playerEnabled:
                  type: boolean
                  default: true
                displayEnabled:
                  type: boolean
                  default: true
                ttlMinutes:
                  type: integer
                  minimum: 15
                  maximum: 1440
      responses:
        "201":
          description: Código generado y código anterior revocado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableAccessGrant"
    delete:
      summary: Revocar el acceso temporal de una campaña
      operationId: revokeTableAccess
      parameters:
        - name: campaignId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Acceso revocado
...
  /worlds:
    get:
      summary: Listar mundos
      operationId: listWorlds
      responses:
        "200":
          description: Lista de mundos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/World"
    post:
      summary: Crear mundo
      operationId: createWorld
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorldRequest"
      responses:
        "201":
          description: Mundo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/World"
  /campaigns:
    get:
      summary: Listar campañas
      operationId: listCampaigns
      responses:
        "200":
          description: Lista de campañas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      summary: Crear campaña
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaña creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /campaigns/{campaignId}/workflow:
    get:
      summary: Consultar lobby, sesiones y presencia de una campaña
      operationId: getSessionWorkflow
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Estado actual del workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWorkflowState"
...
  /campaigns/{campaignId}/ruleset:
    get:
      summary: Obtener ruleset activo de campaña
      operationId: getCampaignRuleset
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Ruleset resuelto para la campaña
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignRulesetResponse"
...
  /campaigns/{campaignId}/assets:
    get:
      summary: Listar assets de campaña
      operationId: listCampaignAssets
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - name: type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AssetLibraryType"
      responses:
        "200":
          description: Biblioteca de assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetLibraryResponse"
    post:
      summary: Registrar asset de campaña
      operationId: createCampaignAsset
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "201":
          description: Asset registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetLibraryItem"
  /campaigns/{campaignId}/assets/{assetId}:
    get:
      summary: Consultar asset de campaña
      operationId: getCampaignAsset
      parameters:
        - $ref: "#/components/parameters/CampaignId"
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetLibraryItem"
        "404":
          description: Asset no encontrado
...
  /campaigns/{campaignId}/encounters:
    post:
      summary: Crear encuentro
      operationId: createEncounter
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEncounterRequest"
      responses:
        "201":
          description: Encuentro creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Encounter"
...
  /campaigns/{campaignId}/npcs:
    get:
      summary: Listar NPCs
      operationId: listNpcs
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Lista de NPCs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NPC"
    post:
      summary: Crear NPC
      operationId: createNpc
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNPCRequest"
      responses:
        "201":
          description: NPC creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NPC"
...
  /campaigns/{campaignId}/enemies:
    get:
      summary: Listar enemigos
      operationId: listEnemies
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      responses:
        "200":
          description: Lista de enemigos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enemy"
    post:
      summary: Crear enemigo
      operationId: createEnemy
      parameters:
        - $ref: "#/components/parameters/CampaignId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnemyRequest"
      responses:
        "201":
          description: Enemigo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enemy"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    dmCookie:
      type: apiKey
      in: cookie
      name: dit_dm_session
    tableCookie:
      type: apiKey
      in: cookie
      name: dit_table_session
  parameters:
    CampaignId:
      name: campaignId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    TableAccessGrant:
      type: object
      required:
        [code, campaignId, playerEnabled, displayEnabled, expiresAt]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9]{6}$"
        campaignId:
          type: string
        sessionId:
          type: string
        playerEnabled:
          type: boolean
        displayEnabled:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    TableAccessStatus:
      type: object
      required: [active, campaignId]
      properties:
        active:
          type: boolean
        campaignId:
          type: string
        sessionId:
          type: string
        playerEnabled:
          type: boolean
        displayEnabled:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    World:
      type: object
      required: [id, name, description, systemTags]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        systemTags:
          type: array
          items:
            type: string
    CreateWorldRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        systemTags:
          type: array
          items:
            type: string
    Campaign:
      type: object
      required: [id, name, ruleset, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        ruleset:
          type: string
          example: dnd5e
        status:
          type: string
          enum: [draft, active, archived]
    CreateCampaignRequest:
      type: object
      required: [name, ruleset]
      properties:
        name:
          type: string
        description:
          type: string
        ruleset:
          type: string
        worldId:
          type: string
        status:
          type: string
          enum: [draft, active, archived]
    GameSession:
      type: object
      required: [id, campaignId, title, phase, summaryPublic, summaryPrivate]
      properties:
        id:
          type: string
        campaignId:
          type: string
        title:
          type: string
        phase:
          type: string
          enum: [preparation, live, ended]
        scheduledAt:
          type: string
        summaryPublic:
          type: string
        summaryPrivate:
          type: string
...
    SessionParticipant:
      type: object
      required: [id, displayName, role, connectedAt]
      properties:
        id:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [dm, player, display]
        connectedAt:
          type: string
          format: date-time
    SessionWorkflowState:
      type: object
      required: [campaign, sessions, participantsBySession]
      properties:
        campaign:
          $ref: "#/components/schemas/Campaign"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/GameSession"
        participantsBySession:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/SessionParticipant"
...
    AssetLibraryType:
      type: string
      enum: [map, token, portrait, music, sound, effect]
    AssetProvider:
      type: string
      enum: [local, remote, generated]
    AssetStatus:
      type: string
      enum: [available, missing, archived]
    AssetLibraryResponse:
      type: object
      required: [version, campaignId, assets, counts, policy, updatedAt]
      properties:
        version:
          type: integer
          const: 1
        campaignId:
          type: string
        assets:
          type: array
          items:
            $ref: "#/components/schemas/AssetLibraryItem"
        counts:
          $ref: "#/components/schemas/AssetLibraryCounts"
        policy:
          $ref: "#/components/schemas/AssetPolicyLimits"
        updatedAt:
          type: string
          format: date-time
    AssetLibraryCounts:
      type: object
      required: [map, token, portrait, music, sound, effect]
      properties:
        map:
          type: integer
        token:
          type: integer
        portrait:
          type: integer
        music:
          type: integer
        sound:
          type: integer
        effect:
          type: integer
    AssetPolicyLimits:
      type: object
      required:
        - maxAssetsPerCampaign
        - maxNameCharacters
        - maxUrlCharacters
        - maxMetadataBytes
        - maxMapDimensionPixels
        - maxAudioDurationSeconds
        - maxDeclaredBytes
      properties:
        maxAssetsPerCampaign:
          type: integer
          example: 500
        maxNameCharacters:
          type: integer
          example: 120
        maxUrlCharacters:
          type: integer
          example: 2048
        maxMetadataBytes:
          type: integer
          example: 16384
        maxMapDimensionPixels:
          type: integer
          example: 8192
        maxAudioDurationSeconds:
          type: integer
          example: 14400
        maxDeclaredBytes:
          type: object
          additionalProperties:
            type: integer
...
    AssetLibraryItem:
      type: object
      required:
        - id
        - campaignId
        - type
        - name
        - url
        - provider
        - status
        - metadata
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        campaignId:
          type: string
        type:
          $ref: "#/components/schemas/AssetLibraryType"
        name:
          type: string
        url:
          type: string
        provider:
          $ref: "#/components/schemas/AssetProvider"
        status:
          $ref: "#/components/schemas/AssetStatus"
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [type, name, url]
      properties:
        version:
          type: integer
          const: 1
        type:
          $ref: "#/components/schemas/AssetLibraryType"
        name:
          type: string
        url:
          type: string
        provider:
          $ref: "#/components/schemas/AssetProvider"
        status:
          $ref: "#/components/schemas/AssetStatus"
        metadata:
          type: object
          additionalProperties: true
        requestId:
          type: string
...
    RulesetSummary:
      type: object
      required: [id, name, version, description, capabilities, tags]
      properties:
        id:
          type: string
          example: dnd5e
        name:
          type: string
        version:
          type: string
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
            enum: [d20-checks, advantage, conditions, initiative, character-sheet]
        tags:
          type: array
          items:
            type: string
    RulesetDefinition:
      allOf:
        - $ref: "#/components/schemas/RulesetSummary"
        - type: object
          required: [dice, combat, characterSheet]
          properties:
            dice:
              type: object
              required: [notationExamples, checks]
              properties:
                notationExamples:
                  type: array
                  items:
                    type: string
                checks:
                  type: array
                  items:
                    $ref: "#/components/schemas/DiceCheckPreset"
            combat:
              type: object
              required: [initiativeFormula, conditions]
              properties:
                initiativeFormula:
                  type: string
                  example: d20+DEX
                conditions:
                  type: array
                  items:
                    $ref: "#/components/schemas/RulesetCondition"
            characterSheet:
              type: object
              required: [sections]
              properties:
                sections:
                  type: array
                  items:
                    $ref: "#/components/schemas/CharacterSheetSection"
    CampaignRulesetResponse:
      type: object
      required:
        - campaignId
        - requestedRulesetId
        - resolvedRulesetId
        - fallbackUsed
        - ruleset
      properties:
        campaignId:
          type: string
        requestedRulesetId:
          type: string
        resolvedRulesetId:
          type: string
        fallbackUsed:
          type: boolean
        ruleset:
          $ref: "#/components/schemas/RulesetDefinition"
    DiceCheckPreset:
      type: object
      required: [id, label, formula, description]
      properties:
        id:
          type: string
        label:
          type: string
        formula:
          type: string
        description:
          type: string
    RulesetCondition:
      type: object
      required: [id, label, publicDescription, rulesHint]
      properties:
        id:
          type: string
        label:
          type: string
        publicDescription:
          type: string
        rulesHint:
          type: string
    CharacterSheetSection:
      type: object
      required: [id, title, fields]
      properties:
        id:
          type: string
        title:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/CharacterSheetField"
    CharacterSheetField:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, number, textarea, select]
        defaultValue:
          oneOf:
            - type: string
            - type: number
        options:
          type: array
          items:
            type: string
...
    CreateEncounterRequest:
      type: object
      required: [sceneId, name, combatants]
      properties:
        version:
          type: integer
          const: 1
        sceneId:
          type: string
        name:
          type: string
        combatants:
          type: array
          items:
            $ref: "#/components/schemas/CreateEncounterCombatant"
    CreateEncounterCombatant:
      type: object
      required: [entityType, name]
      properties:
        tokenId:
          type: string
        entityType:
          type: string
          enum: [player, enemy, npc, object]
        entityId:
          type: string
        name:
          type: string
        initiative:
          type: number
        currentHp:
          type: integer
        temporaryHp:
          type: integer
        conditions:
          type: array
          items:
            type: string
    Encounter:
      type: object
      required: [id, sceneId, status, roundNumber, combatants]
      properties:
        id:
          type: string
        campaignId:
          type: string
        sceneId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [draft, active, completed]
        roundNumber:
          type: integer
        activeCombatantId:
          type: string
        combatants:
          type: array
          items:
            $ref: "#/components/schemas/Combatant"
    Combatant:
      type: object
      required: [id, entityType, name, initiative, turnOrder, conditions]
      properties:
        id:
          type: string
        tokenId:
          type: string
        entityType:
          type: string
          enum: [player, enemy, npc, object]
        name:
          type: string
        initiative:
          type: number
        turnOrder:
          type: integer
        currentHp:
          type: integer
        temporaryHp:
          type: integer
        conditions:
          type: array
          items:
            type: string
...
    NPC:
      type: object
      required: [id, campaignId, name, role, motivation, voice, publicNotes]
      properties:
        id:
          type: string
        campaignId:
          type: string
        name:
          type: string
        role:
          type: string
        motivation:
          type: string
        voice:
          type: string
        publicNotes:
          type: string
        privateNotes:
          type: string
        secrets:
          type: array
          items:
            type: string
    CreateNPCRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        role:
          type: string
        motivation:
          type: string
        voice:
          type: string
        publicNotes:
          type: string
        privateNotes:
          type: string
        secrets:
          type: array
          items:
            type: string
    Enemy:
      type: object
      required: [id, campaignId, name, creatureType, maxHp, currentHp, armorClass]
      properties:
        id:
          type: string
        campaignId:
          type: string
        name:
          type: string
        creatureType:
          type: string
        maxHp:
          type: integer
        currentHp:
          type: integer
        armorClass:
          type: integer
        challengeRating:
          type: string
    CreateEnemyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        creatureType:
          type: string
        maxHp:
          type: integer
        currentHp:
          type: integer
        armorClass:
          type: integer
        challengeRating:
          type: string
...